Revert "fix(sdk): filter media against active model"
This reverts commit bcf1b661641055190757120f56de1ab022864480.
